+++ EasyFinance.Domain.Tests/Financial/ExpenseTests.cs
@@ -10,7 +10,7 @@ namespace EasyFinance.Domain.Tests.Financial
         [Theory]
         [InlineData(-1)]
         [InlineData(-250)]
-        public void AddGoal_SendNegativeGoal_ShouldThrowException(decimal goal)
+        public void AddGoal_SendNegativeGoal_ShouldThrowException(int goal)
         {
             var action = () => new ExpenseBuilder().AddGoal(goal).Build();
